requirements.txt
langchain-ollama>=0.1.0
langchain-core>=0.2.0
PyMuPDF>=1.23.0
Pillow>=10.0.0 